+++ gettext-rs/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 - Documentation on when functions will panic (Alexander Batischev)
+- Functions to query gettext configuration. These are wrappers over
+    `textdomain(NULL)`, `bindtextdomain(domain, NULL)`, and
+    `bind_textdomain_codeset(domain, NULL)` (Alexander Batischev)
 
 ### Changed
 - **Users are required to configure gettext for UTF-8, either by setting a UTF-8